+++ packages/calendar/src/components/calendar-types/travel-date-range-calendar/TravelDateRangeCalendar.tsx
@@ -13,10 +13,11 @@ import {
   DateTextInputVariant,
   TravelDateRangeInputValue,
 } from "../../../features/travel-calendar/types";
+import { SupportedLocaleCode } from "../../../features/localize-date-format/LocaleMapper";
 
 export interface TravelDateRangeCalendarProps
   extends ValueAndOnValueChangeProps<TravelDateRangeInputValue> {
-  localeCode?: string;
+  localeCode?: SupportedLocaleCode;
   initialMonthInFocus?: Date;
   startDateLabel?: string;
   endDateLabel?: string;
@@ -118,6 +119,7 @@ export const TravelDateRangeCalendar: React.FC<
           numMonths={numMonthsInMonthPicker}
           value={visibleMonth}
           size={size}
+          localeCode={localeCode}
           onValueChange={(v) => {
             setVisibleMonth(v);
             setVisiblePanel("calendar");